% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binstall.R
\name{binstall}
\alias{binstall}
\title{Batch Install and Load R Packages}
\usage{
binstall(librarylines)
}
\arguments{
\item{librarylines}{A string containing lines of R code with library() calls.}
}
\description{
This function install and load multiple R packages from a list of library declarations. It extracts
package names from a given set of library() calls, checks for packages that are already installed and
identifies those that are missing. If any packages are missing, it prompt the user for insallation
and loading.
}
