% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binstall.R
\name{buninstall}
\alias{buninstall}
\title{This function uninstalls multiple R packages based on a list of library() declarations. It extracts
package names from the provided lines of R code, checks for packages that are currently installed,
and identifies those that can be uninstalled. If any packages are installed, the function prompts
the user for confirmation before proceeding with the uninstallation.}
\usage{
buninstall(librarylines)
}
\arguments{
\item{librarylines}{A string containing lines of R code with library() calls.}
}
\description{
This function uninstalls multiple R packages based on a list of library() declarations. It extracts
package names from the provided lines of R code, checks for packages that are currently installed,
and identifies those that can be uninstalled. If any packages are installed, the function prompts
the user for confirmation before proceeding with the uninstallation.
}
